fixes to pass the test

+++ sportsclub/sportsclub-domain/src/main/java/org/jboss/snowdrop/samples/sportsclub/domain/repository/EquipmentRepository.java
@@ -9,7 +9,7 @@ import java.util.Date;
 /**
  *
  */
-public interface EquipmentRepository extends Repository<Equipment, Integer>
+public interface EquipmentRepository extends Repository<Equipment, Long>
 {
    public List<Equipment> getAvailableEquipments(Date fromDate, Date toDate);
