+++ sportsclub/sportsclub-reservations-spring/src/main/java/org/jboss/snowdrop/samples/sportsclub/service/EquipmentService.java
@@ -10,5 +10,5 @@ public interface EquipmentService
 {
    public EquipmentType[] getEquipmentTypes();
 
-   public Equipment findEquipmentById(int id);
+   public Equipment findEquipmentById(long id);
 }
